improve the parser structure and remove unnecessary dependencies
the project is tightly coupled to WASM related libraries as one of the
original goals was to use the functionality in the web. It also included
serde and JSON serialization which is overkill for what the purpose of
the library should be.

the structure of the project was extended to better isolate the parsing
logic and break it down into its parts.

tests were also added for each module which are rough around the edges
and should be iterated upon.