create workspace and include cli tool
the library is isolated for parsing puz files, however the need for a
generic cli tool to parse them is more useful in the real world than
forcing the usage of rust and the library itself.

i created the top level workspace and the following workspace packages:

* `/parse` - the parsing library, installed via `puz-parse`
* `/cli` - the cli tool, installed via `puz`

i have also tried to boil down the readme documentation to be more
concise, and included an additional method in the parse library for more
utility.